feat: switch to typescript
