--- .stylelintrc
@@ -1,8 +1,0 @@
-{
-  "extends": "stylelint-config-standard",
-  "rules": {
-    "color-no-invalid-hex": true,
-    "unit-no-unknown": true,
-    "property-no-unknown": true
-  }
-}
