--- public/manifest.json
@@ -1,14 +1,0 @@
-{
-  "short_name": "Festember Campus Ambassador Site",
-  "name": "Festember Campus Ambassador Site",
-  "icons": [
-    {
-      "src": "logo.png",
-      "type": "image/png"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
