+++ src/App.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
+
 import { Routers } from './routes';
 
 export default function App() {
-  return (
-    <Routers />
-  );
+  return <Routers />;
 }
