--- src/components/Home/header.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export function Header() {
-  return (
-    <h1>Festember Crew</h1>
-  );
-}
