--- src/pages/Home.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { Header } from '../components/Home/header';
-
-export function Home() {
-  return (
-    <Header />
-  );
-}
