+++ src/pages/Home.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import Header from '../components/Home/header';
+
+export function Home() {
+  return <Header />;
+}
