+++ src/routes/routes.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { Home } from '../pages/Home';
 
 export const publicRoutes = [
